+++ src/op/mod.rs
@@ -122,6 +122,7 @@ pub mod is_pendant;
 pub mod is_regular;
 pub mod is_simple;
 pub mod is_subdigraph;
+pub mod is_superdigraph;
 pub mod is_symmetric;
 pub mod is_walk;
 pub mod iter_arcs;
@@ -152,6 +153,7 @@ pub use {
     is_regular::IsRegular,
     is_simple::IsSimple,
     is_subdigraph::IsSubdigraph,
+    is_superdigraph::IsSuperdigraph,
     is_symmetric::IsSymmetric,
     is_walk::IsWalk,
     iter_arcs::IterArcs,